consolidated event types into file
