+++ packages/openneuro-app/src/scripts/users/user-notification-accordion-actions.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react"
 import * as Sentry from "@sentry/react"
 import { toast } from "react-toastify"
 import { useMutation } from "@apollo/client"
-import type { MappedNotification } from "../types/user-types"
+import type { MappedNotification } from "../types/event-types"
 import { UPDATE_NOTIFICATION_STATUS_MUTATION } from "../queries/datasetEvents"
 import { StatusActionButton } from "./components/status-action-buttons"
 import iconUnread from "../../assets/icon-unread.png"
