+++ packages/openneuro-app/src/scripts/users/user-notification-accordion.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from "react"
 import { toast } from "react-toastify"
 import * as Sentry from "@sentry/react"
 import styles from "./scss/usernotifications.module.scss"
-import type { MappedNotification } from "../types/user-types"
+import type { MappedNotification } from "../types/event-types"
 import { useMutation, useQuery } from "@apollo/client"
 import { PROCESS_CONTRIBUTOR_REQUEST_MUTATION } from "../queries/datasetEvents.js"
 import { GET_USER } from "../queries/user"
